FAC-517: refactor: rename packages from assessmentuserinvitee to assessmentinvitee
