+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/application/service/assessmentuserinvitee/InviteAssessmentUserService.java
@@ -2,7 +2,7 @@ package org.flickit.assessment.core.application.service.assessmentuserinvitee;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.flickit.assessment.core.application.port.in.assessmentuserinvitee.InviteAssessmentUserUseCase;
+import org.flickit.assessment.core.application.port.in.assessmentinvitee.InviteAssessmentUserUseCase;
 import org.springframework.stereotype.Service;
 
 import java.util.UUID;
