+++ flickit-assessment-core/src/test/java/org/flickit/assessment/core/application/port/in/assessmentinvitee/InviteAssessmentUserUseCaseParamTest.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.core.application.port.in.assessmentuserinvitee;
+package org.flickit.assessment.core.application.port.in.assessmentinvitee;
 
 import jakarta.validation.ConstraintViolationException;
 import org.junit.jupiter.api.Test;
